move generator_nodes under generator
