+++ froql_macro/src/lib.rs
@@ -4,7 +4,6 @@
 extern crate proc_macro;
 
 mod generator;
-mod generator_nodes;
 mod macro_error;
 mod parser;
